padrondl/requirements.txt
beautifulsoup4==4.6.0
bs4==0.0.1
certifi==2017.4.17
chardet==3.0.4
idna==2.5
pkg-resources==0.0.0
progressbar2==3.30.2
python-utils==2.1.0
requests==2.18.1
six==1.10.0
urllib3==1.21.1
